+++ src/hooks/useFetchQuestions.ts
@@ -6,14 +6,13 @@ import { Question } from 'src/types/index';
 export const useFetchQuestions = () => {
   // 質問の配列、ローディング中かどうか、エラーが発生したかどうかをuseStateで管理する
   const [questions, setQuestions] = useState<Question[]>([]);
-  const [isLoading, setIsLoading] = useState<boolean>(false);
+  const [isLoading, setIsLoading] = useState<boolean>(true);
   const [error, setError] = useState<any>(null);
 
   // 質問を取得するためのAPIを呼び出すuseEffect
   useEffect(() => {
     const fetchQuestions = async () => {
       try {
-        setIsLoading(true);
         const response = await axios.get('/api/questions/random');
         setQuestions(response.data);
       } catch (error) {
